+++ cmd/example.go
@@ -47,6 +47,11 @@ var example = `    The following examples make use of a basic.json file which co
 
         cat basic.json | gg -s lev:err -f cal,sta
 
+    Select all logs of two different resource handlers. This is to show the
+    s3object and tccpn resource handler logs together.
+
+        cat basic.json | gg -s obj:8ztu4 -s con:plane -s res:s3obj -s res:tccpn\$ -f res,mes -g loo
+
     Display log messages of drainer resources and format their timestamp using
     to the given format.